Call them "simple targets"
There are two hard problems in computer science, naming, cache
invalidation, and off-by-one errors.  Let's address one of them.

Adding the adjective "simple" to the targets in this proposal emphasizes
that we are focusing on a common or reductive usecase, as opposed to
what it sounds like now: establishing a general framework into which all
C language support will be expected to conform.  Whether we ultimately
do that or not, this proposal does not actually handle the kinds of
features that established C projects actually need, and so to avoid
confusion we talk only about simple targets, and save the other surprise
for another day.

The phrase "simple" is already used several times in this proposal:

> Most projects will not conform to these conventions. However, this is
  expected   of any "simple" convention;

> This gives developers a simple mechanism for "falling back" to C

So the name is a good fit for the language actually used.

The insertion of "simple" leaves open the possibility (whether we
ultimately pursue it or not) that there will be some other kind of
C target handled in some other way.