Explain what a simple target is
The first question with new vocabulary is, what is it?

The motivation section actually already explains the kind of targets this
proposal was designed to address.  I've moved the examples to the front
since they are an immediate way to understand the idea.

The second paragraph contains a more general definition of a simple
target.  They are targets which "build primarily with the Swift package
manager and can be exposed to Swift directly as part of a single
package."  This clearly dilineates the kind of targets the proposal is
intended to address.

This moves the religious question out of scope.  Let's not imply that C
projects *should* build with swiftPM, let's say that *given* they want
to build with swiftPM, how is it accomplished?  It is accomplished by
becoming a simple target.  Say *how* this is done, not that it *should
be.*

We also say that if you don't want to build with swiftPM, you're not a
simple target, and your problems are out of scope.