Clarify what we mean by "we don't think... there is a convention"
Actually there is a straightforward convention to build C targets:
`./configure && make`.  70% of the time, it works every time.

The problem (err, "problem") is that this convention requires external
software, and is kind of a black box in terms of SwiftPM's understanding
of the source graph, and we don't want to encourage new software to
build this way.  So we include the phrase "for directly building C
targets in the Swift package manager" to qualify our "no convention"
claim.